fix(budget): use u128 arithmetic for texture soft-limit comparison [hud-nyrss]
Prevent theoretical u64 overflow in check_budget_soft when
texture_bytes_used or max_texture_bytes are near u64::MAX.
The session path caps at HARD_MAX_TEXTURE_BYTES (2 GiB) so overflow
cannot occur in production, but the function signature carries no
such precondition and tests construct uncapped budgets directly.
Cast both operands to u128 before the *5/*4 comparison.

+++ crates/tze_hud_scene/src/lease/budget.rs
@@ -121,7 +121,7 @@ pub fn check_budget_soft(
         idx += 1;
     }
 
-    if usage.texture_bytes_used * 5 >= budget.max_texture_bytes.max(1) * 4 {
+    if (usage.texture_bytes_used as u128) * 5 >= (budget.max_texture_bytes.max(1) as u128) * 4 {
         result[idx] = Some(BudgetDimension::TextureBytes);
         idx += 1;
     }
